+++ src/utest/process.cpp
@@ -6,7 +6,7 @@
 
 static void* myalloc(void*, void* old, aint_t sz)
 {
-    return realloc(old, sz);
+    return realloc(old, (size_t)sz);
 }
 
 TEST_CASE("process_allocation")
